+++ depthimage_to_pointcloud2/include/depthimage_to_pointcloud2/depth_traits.hpp
@@ -33,14 +33,14 @@
 // This file is originally from:
 // https://github.com/ros-perception/image_pipeline/blob/da750d1/depth_image_proc/include/depth_image_proc/depth_traits.h  // NOLINT
 
-#ifndef DEPTH_TO_POINTCLOUD__DEPTH_TRAITS_HPP_
-#define DEPTH_TO_POINTCLOUD__DEPTH_TRAITS_HPP_
+#ifndef DEPTHIMAGE_TO_POINTCLOUD2__DEPTH_TRAITS_HPP_
+#define DEPTHIMAGE_TO_POINTCLOUD2__DEPTH_TRAITS_HPP_
 
 #include <algorithm>
 #include <limits>
 #include <vector>
 
-namespace depth_to_pointcloud
+namespace depthimage_to_pointcloud2
 {
 
 // Encapsulate differences between processing float and uint16_t depths
@@ -71,6 +71,6 @@ struct DepthTraits<float>
   }
 };
 
-}  // namespace depth_to_pointcloud
+}  // namespace depthimage_to_pointcloud2
 
-#endif  // DEPTH_TO_POINTCLOUD__DEPTH_TRAITS_HPP_
+#endif  // DEPTHIMAGE_TO_POINTCLOUD2__DEPTH_TRAITS_HPP_
