+++ depthimage_to_pointcloud2/package.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0"?>
 <package format="2">
-  <name>depth_to_pointcloud</name>
+  <name>depthimage_to_pointcloud2</name>
   <version>0.0.2</version>
   <description>A simple node to convert a depth image and camera info into a PointCloud2.</description>
   <maintainer email="clalancette@osrfoundation.org">Chris Lalancette</maintainer>
